+++ movie-app/src/components/DeleteCartItem.js
@@ -2,33 +2,34 @@ import {Trash3Fill} from "react-bootstrap-icons";
 import {useCart} from "../contexts/CartContext";
 import axios from "axios";
 import toastify from "../consts/toastify";
-import {DELETE_FROM_CART_FAIL_MSG} from "../consts/consts";
+import {DELETE_FROM_CART_FAIL_MSG, Action} from "../consts/consts";
 import {useState} from "react";
 import Spinner from "./Spinner";
 
 export default function DeleteCartItem({id}){
-    const [isLoad, setIsLoad] = useState(false);
+    const [isLoading, setIsLoading] = useState(false);
     const {dispatch} = useCart();
 
     //TODO: move to usefetch
     const deleteItem = async (id) =>{
-        setIsLoad(true)
+        if(isLoading)return;
+        setIsLoading(true)
         try{
             const res = await axios.delete('/api/cart/' + id);
             console.log("deleted " ,id, res);
-            dispatch({type:'delete', payload: {id}});
+            dispatch({type: Action.DELETE, payload: {id}});
         }
         catch (err){
             toastify.errorToast(DELETE_FROM_CART_FAIL_MSG);
         }
         finally {
-            setIsLoad(false)
+            setIsLoading(false)
         }
     }
 
     return (
         <button onClick={()=>deleteItem(id)} className="btn btn-light">
-            {isLoad ? <Spinner/> : <Trash3Fill/>}
+            {isLoading ? <Spinner/> : <Trash3Fill/>}
         </button>
     )
 }
